pyproject.toml
[project]
name = "pythonproject"
version = "0.1.0"
description = ""
authors = [
    {name = "calamistratus",email = "maksgalk45@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "redis (>=7.1.0,<8.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "fastapi (>=0.121.3,<0.122.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
